pyproject.toml
[tool.poetry]
name = "blackbox-fetch"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
PyQt6 = "^6.2.0"
pyserial = "^3.5"
yamspy = "^0.3.3"

[tool.poetry.dev-dependencies]
pyinstaller = "^4.5.1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
